pyproject.toml
[tool.poetry]
name = "mlops-23-autumn"
version = "0.1.0"
description = "Project for my tasks in mlops course"
authors = ["Ilya Kochagin <ilyakochagin@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8,<3.12"
absl-py = "2.0.0"
astunparse = "1.6.3"
cachetools = "5.3.1"
certifi = "2023.7.22"
charset-normalizer = "3.3.0"
contourpy = "1.1.1"
cycler = "0.12.0"
filelock = "3.12.4"
flatbuffers = "23.5.26"
fonttools = "4.43.1"
fsspec = "2023.9.2"
gast = "0.4.0"
google-auth = "2.23.2"
google-auth-oauthlib = "1.0.0"
google-pasta = "0.2.0"
grpcio = "1.59.0"
h5py = "3.9.0"
idna = "3.4"
importlib-metadata = "6.8.0"
importlib-resources = "6.1.0"
jinja2 = "3.1.2"
joblib = "1.3.2"
keras = "2.13.1"
kiwisolver = "1.4.5"
libclang = "16.0.6"
markdown = "3.4.4"
markupsafe = "2.1.3"
matplotlib = "3.7.3"
mlxtend = "0.23.0"
mpmath = "1.3.0"
networkx = "3.1"
numpy = "1.24.3"
oauthlib = "3.2.2"
opt-einsum = "3.3.0"
packaging = "23.2"
pandas = "2.0.3"
pillow = "10.0.1"
protobuf = "4.24.4"
pyasn1 = "0.5.0"
pyasn1-modules = "0.3.0"
pyparsing = "3.1.1"
python-dateutil = "2.8.2"
pytz = "2023.3.post1"
requests = "2.31.0"
requests-oauthlib = "1.3.1"
rsa = "4.9"
scikit-learn = "1.3.1"
scipy = "1.10.1"
seaborn = "0.13.0"
six = "1.16.0"
sympy = "1.12"
termcolor = "2.3.0"
threadpoolctl = "3.2.0"
torch = "2.1.0"
typing-extensions = "4.5.0"
tzdata = "2023.3"
urllib3 = "2.0.6"
werkzeug = "3.0.0"
wrapt = "1.15.0"
zipp = "3.17.0"
pyyaml = "6.0.1"
setuptools-scm = "8.0.4"
tomli = "2.0.1"
cfgv = "3.4.0"
distlib = "0.3.7"
identify = "2.5.30"
nodeenv = "1.8.0"
platformdirs = "3.11.0"
pre-commit = "3.4.0"
virtualenv = "20.24.5"
dvc = "^3.30.3"
dvc-gdrive = "^2.20.0"
hydra-core = "^1.3.2"
onnx = "^1.15.0"
tritonclient = "^2.41.0"
gevent = "^23.9.1"
geventhttpclient = "^2.0.11"

[tool.black]
line-length = 90
target-version = ["py38"]

[tool.isort]
src_paths = ["ds_project"]
profile = "black"
line_length = 90
lines_after_imports = 2

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
